Moved annotation-related Javascript to be bundled with webpack
