+++ app/assets/javascripts/Results/main.js
@@ -1,15 +1,3 @@
-// Source code highlighting
-//= require Annotations/globals
-//= require Annotations/annotation_manager
-//= require Annotations/annotation_text
-//= require Annotations/annotation_text_displayer
-//= require Annotations/annotation_text_manager
-//= require Annotations/source_code_line
-//= require Annotations/text_annotation_manager
-//= require Annotations/image_annotation_manager
-//= require Annotations/pdf_annotation_manager
-//= require Annotations/html_annotations
-
 // Page-specific Javascript
 //= require Grader/marking
 //= require panes
