+++ app/javascript/common/annotations/globals.js
@@ -1,8 +1,8 @@
 // Global variables currently used by the annotation code
 // TODO: remove these!
-var annotation_manager = null;
+window.annotation_manager = null;
 
-var ANNOTATION_TYPES = {
+window.ANNOTATION_TYPES = {
   CODE: 0,
   IMAGE: 1,
   PDF: 2,
@@ -11,4 +11,4 @@ var ANNOTATION_TYPES = {
 
 // Enum to tell the code if an image, code, or pdf is being shown
 // in the codeviewer
-var annotation_type;
+window.annotation_type = null;
