--- app/views/results/common/_syntax_highlighter_brushes.html.erb
@@ -1,12 +1,0 @@
-<%# SOURCE CODE GLOWER AND IMAGE ANNOTATION MANAGER %>
-<%= javascript_include_tag 'Annotations/globals',
-                           'Annotations/annotation_manager',
-                           'Annotations/annotation_text',
-                           'Annotations/annotation_text_displayer',
-                           'Annotations/AnnotationTextManager',
-                           'Annotations/source_code_line',
-                           'Annotations/text_annotation_manager',
-                           'Annotations/image_annotation_manager',
-                           'Annotations/pdf_annotation_manager',
-                           'Annotations/html_annotations',
-                           nonce: true %>
